Logic and UI commit
I ve added 2 function to Game class. One is for switching the current player and the second one is checking where at which index do we set a tag. in GameViewController I have a func onPress that takes that function created in Game class connects it with a UI tag and sets a symbol for it.   Func setSymbol checks who is the currentplayer and based on that it sets a corresponded image.