PlayerVSComputer
the game has gameOption to play with computer and that is set up in the game logic. right now the game can be played but the computer is overlapping spots that are already filled. todo: crate a function to check so the computer does not place his move on taken spots